durable-image-pipeline/requirements.txt
azure-functions
azure-functions-durable
pillow
pyodbc>=4.0.32

